encrypted file upload fixed

+++ pg-backup/src/Encryption.ts
@@ -5,10 +5,13 @@ import { existsSync } from "fs";
 
 export default class Encryption {
 
-    static async encryptFile(filePath: string) {
+    static async encryptFile(filePath: string, deleteFile = true) {
 
         const encPath = `${filePath}.enc`;
         await spawnPromise("openssl", ["enc", "-aes-256-cbc", "-pbkdf2", "-in", filePath, "-out", encPath, "-pass", "file:/app/.pwd-hash" ]);
+        if (!deleteFile) {
+            return;
+        }
         await unlink(filePath);
     }
